# kisaan-backend-node/openapi.yaml
openapi: 3.0.3
info:
  title: KisaanCenter API
  version: 1.0.0
  description: Full API documentation for KisaanCenter backend
servers:
  - url: http://localhost:3000/api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Settlement:
      type: object
      properties:
        id: { type: integer }
        shop_id: { type: integer }
        user_id: { type: integer }
        amount: { type: number }
        reason: { type: string, enum: [overpayment, underpayment, adjustment] }
        status: { type: string, enum: [pending, settled], default: pending }
        settlement_date: { type: string, format: date-time, nullable: true }
        notes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [shop_id, user_id, amount, reason]
    SettlementCreate:
      type: object
      properties:
        shop_id: { type: integer }
        user_id: { type: integer }
        amount: { type: number }
        reason: { type: string, enum: [overpayment, underpayment, adjustment] }
        notes: { type: string, nullable: true }
      required: [shop_id, user_id, amount, reason]
...
    PaginatedResponse:
      type: object
      properties:
        data: { type: array }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }
    ApiResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: { }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, default: false }
        message: { type: string }
        error: { type: string }
...
paths:
...
  /settlements:
    post:
      summary: Create settlement
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SettlementCreate' }
      responses:
        '201':
          description: Settlement created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Settlement' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    get:
      summary: Get all settlements
      parameters:
        - in: query
          name: shop_id
          schema: { type: integer }
        - in: query
          name: user_id
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string, enum: [pending, settled] }
        - in: query
          name: reason
          schema: { type: string, enum: [overpayment, underpayment, adjustment] }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      responses:
        '200':
          description: List of settlements
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Settlement' }
  /settlements/{id}:
    get:
      summary: Get settlement by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Settlement found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Settlement' }
        '404':
          description: Settlement not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    put:
      summary: Update settlement
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [pending, settled] }
                settlement_date: { type: string, format: date-time }
                notes: { type: string }
      responses:
        '200':
          description: Settlement updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Settlement' }
        '404':
          description: Settlement not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /settlements/settle/{settlement_id}:
    post:
      summary: Settle amount
      parameters:
        - in: path
          name: settlement_id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Amount settled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /settlements/expense:
    post:
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shop_id: { type: integer }
                amount: { type: number, minimum: 0 }
                description: { type: string }
              required: [shop_id, amount, description]
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /credits:
    post:
      summary: Create credit advance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id: { type: integer }
                amount: { type: number, minimum: 0 }
                reason: { type: string }
              required: [user_id, amount]
      responses:
        '201':
          description: Credit advance created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
    get:
      summary: Get credit advances
      parameters:
        - in: query
          name: user_id
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string, enum: [pending, approved, rejected] }
      responses:
        '200':
          description: List of credit advances
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: integer }
                            user_id: { type: integer }
                            amount: { type: number }
                            status: { type: string }
                            reason: { type: string }
                            created_at: { type: string, format: date-time }
...
  /commissions:
    get:
      summary: Get commissions
      parameters:
        - in: query
          name: shop_id
          schema: { type: integer }
        - in: query
          name: from_date
          schema: { type: string, format: date }
        - in: query
          name: to_date
          schema: { type: string, format: date }
      responses:
        '200':
          description: List of commissions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: integer }
                            shop_id: { type: integer }
                            amount: { type: number }
                            transaction_id: { type: integer }
                            created_at: { type: string, format: date-time }
    post:
      summary: Create commission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shop_id: { type: integer }
                rate: { type: number, minimum: 0 }
                type: { type: string, enum: [percentage, fixed] }
              required: [shop_id, rate, type]
      responses:
        '201':
          description: Commission created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...